pyproject.toml
[project]
name = "django-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "split-settings>=1.0.0",
]
